Initital CLI client to fetch collection
The cli is written with the `cobra` package
and can be invoked via the `outline` command.

For now, we support a `collection` subcommand
to operate on anything related outline collections.

The `fetch` subcommand on the `collection` command can be
triggered to print a specific collection in json format to stdout.