Add code linting to the workflow (#22)
Code linting jobs are missing in CI so here we add them. The golangci-
lint is a well known tool for the job and for now we use it in its
default configuration. It can be configured more if needed.
Reference: https://github.com/golangci/golangci-lint-action

We also rearranged the existing test workflow to be consistent in file
structure and job naming.