New workflow for drafting releases with archives
Running the new workflow for a given tag should allows us to draft new
release for the tag. The draft also uploads app archives containing app
executables for Linux and MacOS platoforms plus some metadata.

To achieve this we the https://goreleaser.com/ project.

We also add a new sub-command 'version' which obviously prints the git
tag from which the app binary was built.